Fixed some range-check errors. Explicitly disable range-checks {$R-} should not be necessary for the MacOSX build anymore. Note: Texture-numbers (e.g. texnum) are unsigned (GLuint) so do not init them to -1. Use 0 instead.
git-svn-id: https://svn.code.sf.net/p/ultrastardx/svn/trunk@1040 b956fd51-792f-4845-bead-9b4dfca2ff2c

+++ Game/Code/Classes/UCommon.pas
@@ -387,7 +387,7 @@ begin
   end;
   // create window
   Result := CreateWindowEx(WS_EX_TOOLWINDOW, WndClass.lpszClassName, '',
-    WS_POPUP, 0, 0, 0, 0, 0, 0, HInstance, nil);
+    DWORD(WS_POPUP), 0, 0, 0, 0, 0, 0, HInstance, nil);
   if (Result = 0) then
     Exit;
   // assign individual callback procedure to the window
